requirements.txt
Pillow
playwright
htmldocx
python-docx
vtracer
